Fix path passed to do_metta/5 for load_metta code action
This makes the "evaluate metta" code action actually show the value of
the evaluated form.

Wrapping the path in lsp(...) seems to result in no output. Digging in
to metta_repl.pl, I see u_do_metta_exec000 checks to see if the exec
is coming from the LSP, but just seems to execute the input anyway,
albeit in a way that seems not to do what I want.