[lsp] Handle VSCode-style TCP client starting
Very strangely, when VSCode spawns the process using a socket
connection, the *editor* starts the server that the language server
connects to as a client. Apparently to avoid race conditions.

See https://github.com/microsoft/language-server-protocol/issues/604#issuecomment-439853638
(the discussion is about the IPC mechanism, but linked comment is true
for TCP as well)