Fix debug flag setting for with_no_debug/1
The built-in prolog flag `debug` is of type `bool`, so setting it to
`silent` causes an error when create_prolog_flag is called. That error
doesn't seem to appear always, but when running in another thread, I
can see the error getting printed out. e.g. running the below shows
the error in the top-level.

```prolog
thread_create((
   catch(
    create_prolog_flag(debug, silent,
         [keep(false), access(read_write), type(term)]),
       E,
       format(user_error, "ERR SETTING ~q~n", [E])),
   set_prolog_flag(debug, silent) ), _).
```

This causes a problem in the LSP hover, as occasionally the error
output gets interleaved with the displayed content